+++ examples/009-clean_architecture_todo_mvc_mutable_state/lib/main.dart
@@ -3,23 +3,9 @@
 // in the LICENSE file.
 
 import 'package:flutter/material.dart';
-import 'package:shared_preferences/shared_preferences.dart';
-import 'package:states_rebuilder/states_rebuilder.dart';
 
 import 'app.dart';
-import 'data_source/todo_repository.dart';
 
 void main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  runApp(
-    Injector(
-        inject: [
-          Inject<SharedPreferences>.future(
-            () => SharedPreferences.getInstance(),
-          ),
-        ],
-        builder: (context) {
-          return StatesRebuilderApp();
-        }),
-  );
+  runApp(StatesRebuilderApp());
 }