+++ states_rebuilder_package/lib/src/states_rebuilder_debug.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/widgets.dart';
-import 'injector.dart';
 
+import 'injector.dart';
 import 'states_rebuilder.dart';
 
 class StatesRebuilderDebug {