+++ states_rebuilder_package/lib/src/when_connection_state.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
-import '../states_rebuilder.dart';
+
 import 'reactive_model.dart';
+import 'rm_key.dart';
 import 'state_builder.dart';
 
 ///a combination of [StateBuilder] widget and [ReactiveModel.whenConnectionState] method.