+++ states_rebuilder_package/lib/src/when_rebuilder_or.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import '../states_rebuilder.dart';
 import 'reactive_model.dart';
+import 'rm_key.dart';
 import 'state_builder.dart';
 import 'when_connection_state.dart';